Move click helper func to utils
