+++ dwave/cloud/utils.py
@@ -1,5 +1,7 @@
 from __future__ import division, absolute_import
 
+from functools import wraps
+
 import six
 import readline
 
@@ -63,3 +65,31 @@ def readline_input(prompt, prefill=''):
         return six.moves.input(prompt)
     finally:
         readline.set_startup_hook()
+
+
+def click_info_switch(f):
+    """Decorator to create eager Click info switch option, as described in:
+    http://click.pocoo.org/6/options/#callbacks-and-eager-options.
+
+    Takes a no-argument function and abstracts the boilerplate required by
+    Click (value checking, exit on done).
+
+    Example:
+
+        @click.option('--my-option', is_flag=True, callback=my_option,
+                    expose_value=False, is_eager=True)
+        def test():
+            pass
+
+        @click_info_switch
+        def my_option()
+            click.echo('some info related to my switch')
+    """
+
+    @wraps(f)
+    def wrapped(ctx, param, value):
+        if not value or ctx.resilient_parsing:
+            return
+        f()
+        ctx.exit()
+    return wrapped
